agents/requirements.txt
# Core agent framework
langgraph==1.0.0
langchain==0.1.0
langchain-openai==0.1.0

# LLM routing and proxy
litellm==1.30.0
langfuse==2.0.0

# Browser automation
browserless==1.0.0
playwright==1.40.0
websockets==12.0

# Data storage and databases
minio==7.2.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# PDF report generation
reportlab==4.1.0

# Utilities
pydantic==2.5.0
python-dotenv==1.0.0
aiohttp==3.9.1
requests==2.31.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
